feat(server): Re-wrote the handle Request Reference logic
From the ideas suggested and mentioned during the elaboration of how we wanted to cluster the graduate reference on request in the form of array of objects and then save those request into the database as an individual object simultaneously, I have made some changes to the logic to fit the demands as requested. I hope this satisfies the problem at hand. Please leave me a comment where due.